requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.6
drf-yasg==1.21.7

